requirements.txt
absl-py==0.9.0
certifi==2020.6.20
chardet==3.0.4
deepdiff==5.0.2
enum34==1.1.10
future==0.18.2
idna==2.10
mock==3.0.5
mpyq==0.2.5
numpy==1.18.5
ordered-set==4.0.2
portpicker==1.3.1
protobuf==3.12.4
pygame==1.9.6
PySC2==3.0.0
requests==2.24.0
s2clientprotocol==5.0.2.81102.0
s2protocol==5.0.1.81009.0
scipy==1.4.1
six==1.15.0
sk-video==1.1.10
urllib3==1.25.10
websocket-client==0.57.0
whichcraft==0.6.1
